+++ Changelog.rst
@@ -6,14 +6,16 @@ dev
 
 **New**
 
-- Add tests for `Rect`, `Point`
+- Add tests
 - Add `__radd__`, `__rsub__` and `__neg__` for Point.
+- Add `inverse` option for remaining threhsolding functions
 
 **Changes**
 
 - `Rect.__contains__` now acts as `Rect` in C++ with being inclusive on the
   top-left side, and exclusive on the bottom-right side.
 - `Rect` must now have a positive width and height.
+- Resize parameters factor and shape changed to keyword only
 
 
 v0.2.1 (2019-04-05)